Move TextUtils-dependent Utils tests to instrumentation
